Comments: move comment management to new module CommentsHelper.
This removes a nasty dependency of the data layer on the display layer (!)
since CommentsPanel was the previous manager.
