+++ mmstudio/src/main/java/org/micromanager/display/internal/inspector/CommentsPanel.java
@@ -46,6 +46,7 @@ import org.micromanager.data.Annotation;
 import org.micromanager.data.Coords;
 import org.micromanager.data.Datastore;
 import org.micromanager.data.Image;
+import org.micromanager.data.internal.CommentsHelper;
 import org.micromanager.data.DatastoreRewriteException;
 import org.micromanager.data.Image;
 import org.micromanager.data.Metadata;
@@ -63,18 +64,12 @@ import org.micromanager.internal.utils.ReportingUtils;
 
 
 public class CommentsPanel extends InspectorPanel {
-   /** File that comments are saved in. */
-   private static final String COMMENTS_FILE = "comments.txt";
-   /** String key used to access comments in annotations. */
-   private static final String COMMENTS_KEY = "comments";
-
    private JTextArea imageCommentsTextArea_;
    private JTextArea summaryCommentsTextArea_;
    private boolean shouldIgnoreUpdates_ = false;
    private JLabel errorLabel_;
    private DataViewer display_;
    private Datastore store_;
-   private Annotation annotation_;
    private Thread updateThread_;
    private boolean shouldShowUpdates_ = true;
    private Image curImage_ = null;
@@ -147,30 +142,15 @@ public class CommentsPanel extends InspectorPanel {
          // We're in the middle of manually updating the text fields.
          return;
       }
-      if (annotation_ == null) {
-         // Unable to load annotation.
-         return;
-      }
       Coords imageCoords = display_.getDisplayedImages().get(0).getCoords();
       // Determine if anything has actually changed.
       String imageText = imageCommentsTextArea_.getText();
       String summaryText = summaryCommentsTextArea_.getText();
 
-      PropertyMap imageProps = annotation_.getImageAnnotation(imageCoords);
-      if (imageProps == null) {
-         imageProps = MMStudio.getInstance().data().getPropertyMapBuilder().build();
-      }
-      imageProps = imageProps.copy().putString(COMMENTS_KEY, imageText).build();
-      annotation_.setImageAnnotation(imageCoords, imageProps);
-
-      PropertyMap generalProps = annotation_.getGeneralAnnotation();
-      if (generalProps == null) {
-         generalProps = MMStudio.getInstance().data().getPropertyMapBuilder().build();
-      }
-      generalProps = generalProps.copy().putString(COMMENTS_KEY, summaryText).build();
-      annotation_.setGeneralAnnotation(generalProps);
+      CommentsHelper.setImageComment(store_, imageCoords, imageText);
+      CommentsHelper.setSummaryComment(store_, summaryText);
       try {
-         annotation_.save();
+         CommentsHelper.saveComments(store_);
       }
       catch (IOException e) {
          errorLabel_.setText("Error writing comments to disk.");
@@ -236,24 +216,18 @@ public class CommentsPanel extends InspectorPanel {
       if (image == curImage_) {
          return;
       }
-      if (annotation_ == null) {
-         // Unable to get annotation.
-         return;
-      }
       curImage_ = image;
       Coords coords = image.getCoords();
-      if (annotation_.getImageAnnotation(coords) != null) {
-         final String newComments = annotation_.getImageAnnotation(coords).getString(COMMENTS_KEY, "");
-         // Run this in the EDT.
-         SwingUtilities.invokeLater(new Runnable() {
-            @Override
-            public void run() {
-               shouldIgnoreUpdates_ = true;
-               imageCommentsTextArea_.setText(newComments);
-               shouldIgnoreUpdates_ = false;
-            }
-         });
-      }
+      final String newComments = CommentsHelper.getImageComment(store_, coords);
+      // Run this in the EDT.
+      SwingUtilities.invokeLater(new Runnable() {
+         @Override
+         public void run() {
+            shouldIgnoreUpdates_ = true;
+            imageCommentsTextArea_.setText(newComments);
+            shouldIgnoreUpdates_ = false;
+         }
+      });
    }
 
    @Subscribe
@@ -285,17 +259,9 @@ public class CommentsPanel extends InspectorPanel {
       }
       store_ = display_.getDatastore();
 
-      try {
-         annotation_ = store_.loadAnnotation(COMMENTS_FILE);
-      }
-      catch (IOException e) {
-         errorLabel_.setText("Unable to load comments file.");
-      }
-
       shouldIgnoreUpdates_ = true;
-      if (annotation_.getGeneralAnnotation() != null) {
-         summaryCommentsTextArea_.setText(annotation_.getGeneralAnnotation().getString(COMMENTS_KEY, ""));
-      }
+      summaryCommentsTextArea_.setText(
+            CommentsHelper.getSummaryComment(store_));
       shouldIgnoreUpdates_ = false;
       display_.registerForEvents(this);
       store_.registerForEvents(this);
@@ -322,111 +288,5 @@ public class CommentsPanel extends InspectorPanel {
       }
       shouldShowUpdates_ = false;
       updateThread_.interrupt(); // It will then close.
-   }
-
-   /**
-    * Returns the summary comment for the specified Datastore, or "" if it
-    * does not exist.
-    */
-   public static String getSummaryComment(Datastore store) {
-      if (!store.hasAnnotation(COMMENTS_FILE)) {
-         return "";
-      }
-      try {
-         Annotation annotation = store.loadAnnotation(COMMENTS_FILE);
-         if (annotation.getGeneralAnnotation() == null) {
-            return "";
-         }
-         return annotation.getGeneralAnnotation().getString(COMMENTS_KEY, "");
-      }
-      catch (IOException e) {
-         ReportingUtils.logError(e, "Error accessing comments annotation");
-         return "";
-      }
-   }
-
-   /**
-    * Write a new summary comment for the given Datastore.
-    */
-   public static void setSummaryComment(Datastore store, String comment) {
-      try {
-         Annotation annotation = store.loadAnnotation(COMMENTS_FILE);
-         PropertyMap prop = annotation.getGeneralAnnotation();
-         if (prop == null) {
-            prop = MMStudio.getInstance().data().getPropertyMapBuilder().build();
-         }
-         prop = prop.copy().putString(COMMENTS_KEY, comment).build();
-         annotation.setGeneralAnnotation(prop);
-      }
-      catch (IOException e) {
-         ReportingUtils.logError(e, "Error setting summary comment");
-      }
-   }
-
-   /**
-    * Returns the comment for the specified Image in the specified Datastore,
-    * or "" if it does not exist.
-    */
-   public static String getImageComment(Datastore store, Coords coords) {
-      if (!store.hasAnnotation(COMMENTS_FILE)) {
-         return "";
-      }
-      try {
-         Annotation annotation = store.loadAnnotation(COMMENTS_FILE);
-         if (annotation.getImageAnnotation(coords) == null) {
-            return "";
-         }
-         return annotation.getImageAnnotation(coords).getString(COMMENTS_KEY, "");
-      }
-      catch (IOException e) {
-         ReportingUtils.logError(e, "Error accessing comments annotation");
-         return "";
-      }
-   }
-
-   /**
-    * Write a new image comment for the given Datastore.
-    */
-   public static void setImageComment(Datastore store, Coords coords,
-         String comment) {
-      try {
-         Annotation annotation = store.loadAnnotation(COMMENTS_FILE);
-         PropertyMap prop = annotation.getImageAnnotation(coords);
-         if (prop == null) {
-            prop = MMStudio.getInstance().data().getPropertyMapBuilder().build();
-         }
-         prop = prop.copy().putString(COMMENTS_KEY, comment).build();
-         annotation.setImageAnnotation(coords, prop);
-      }
-      catch (IOException e) {
-         ReportingUtils.logError(e, "Error setting image comment");
-      }
-   }
-
-   /**
-    * Return true if there's a comments annotation.
-    */
-   public static boolean hasAnnotation(Datastore store) {
-      return store.hasAnnotation(COMMENTS_FILE);
-   }
-
-   /**
-    * Create a new comments annotation.
-    */
-   public static void createAnnotation(Datastore store) {
-      store.createNewAnnotation(COMMENTS_FILE);
-   }
-
-   /**
-    * Save the store's comments annotation.
-    */
-   public static void save(Datastore store) {
-      try {
-         Annotation annotation = store.loadAnnotation(COMMENTS_FILE);
-         annotation.save();
-      }
-      catch (IOException e) {
-         ReportingUtils.logError(e, "Error saving comment annotations");
-      }
    }
 }
